fix(hid): stop drain loops spinning when the device stops responding
Same defect as the CorsairPeripheralV2 drain loop in the previous commit.
hidapi returns int and yields -1 on error, but these callers stored it in
a uint8_t, so -1 became 255 and their loop conditions stayed true forever.
A device that stops responding pins a core.

AsusAuraCoreLaptopController::GetKeyboardLayout carries two. The drain
loop at the top spins. Below it, hid_get_feature_report failing also gave
255, so `if(result > 0)` returned rd_buf[12], which the memset above
leaves as 0x00, instead of falling through to the ANSI (0x01) fallback
the function already has. Both values select KEYBOARD_LAYOUT_ANSI_QWERTY
at the only caller, so that half is a tidy-up rather than a fix.

The three SteelSeries Aerox GetFirmwareVersion loops read

    while(result > 0 && buffer[0] != CMD)

and hidapi leaves buffer untouched on error, so buffer[0] stays at its
initialised 0x00 and never matches CMD (0x90). Both halves stay true.

These all run during detection, so a failing device hangs detection
rather than burning a core in the background.

Not touched here: the NZXT Hue1/Hue2 and Cryorig H7 read loops spin on
the same disconnect, but they store hid_read's return in an unsigned int
and loop until it equals an expected length (21, 64 and 17 respectively)
with no error check at all. A signed type does not fix those, since -1
equals none of those lengths either way. They need a real error path.

+++ Controllers/AsusAuraCoreController/AsusAuraCoreLaptopController/AsusAuraCoreLaptopController.cpp
@@ -98,7 +98,7 @@ std::string AsusAuraCoreLaptopController::GetDeviceDescription()
 unsigned int AsusAuraCoreLaptopController::GetKeyboardLayout()
 {
     const uint8_t index                                     = 6;
-    uint8_t result                                          = 0;
+    int     result                                          = 0;
     uint8_t rd_buf[ASUSAURACORELAPTOP_WRITE_PACKET_SIZE]    = { ASUSAURACORELAPTOP_REPORT_ID };
     uint8_t buffer[ASUSAURACORELAPTOP_WRITE_PACKET_SIZE]    = { ASUSAURACORELAPTOP_REPORT_ID,
                                                                 ASUSAURACORELAPTOP_CMD_LAYOUT,
